requirements.txt
streamlit==1.40.1
plotly==5.24.1
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.0